# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GPL 3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        }
    },
    "paths": {
...
        "/guest/delete": {
            "delete": {
                "security": [
                    {
                        "GuestAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guest"
                ],
                "summary": "Delete a guest token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/guest/new": {
            "post": {
                "description": "Generate a guest token for anonymous access.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guest"
                ],
                "summary": "Generate a guest token",
                "parameters": [
                    {
                        "description": "Guest details",
                        "name": "guest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.NewGuestReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/DTO.GuestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/guest/scan/{id}": {
            "post": {
                "security": [
                    {
                        "GuestAuth": []
                    }
                ],
                "description": "Scan a story using a guest token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "story",
                    "guest"
                ],
                "summary": "Scan a story",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Story ID to be scanned",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Guest token for authentication",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/guest/stories": {
            "get": {
                "security": [
                    {
                        "GuestAuth": []
                    }
                ],
                "description": "Retrieve a story feed based on specified options.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "story",
                    "guest"
                ],
                "summary": "Get story feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sort the stories by a field (e.g., 'created')",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of returned stories",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset the returned stories",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter stories by start date (e.g.: 2006-01-02T15:04:05)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter stories by end date (e.g: 2006-01-02T15:04:05)",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Guest token for authentication",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.StoriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/guest/verify": {
            "post": {
                "description": "Check the validity of a guest token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guest"
                ],
                "summary": "Check a guest token",
                "parameters": [
                    {
                        "description": "Guest token to be checked",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.VerifyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/DTO.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DTO.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "DTO.GuestResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                },
                "userAgent": {
                    "type": "string",
                    "example": "Mozilla/5.0"
                }
            }
        },
...
        "DTO.NewGuestReq": {
            "type": "object",
            "properties": {
                "displayDetails": {
                    "type": "string",
                    "example": "1920x1080"
                },
                "operatingSystem": {
                    "type": "string",
                    "example": "Linux"
                },
                "versionNumber": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "DTO.Status": {
            "type": "string",
            "enum": [
                "private",
                "public"
            ],
            "x-enum-varnames": [
                "StatusPrivate",
                "StatusPublic"
            ]
        },
        "DTO.StoriesResponse": {
            "type": "object",
            "properties": {
                "stories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.StoryResponse"
                    }
                }
            }
        },
...
        "DTO.StoryResponse": {
            "type": "object",
            "properties": {
                "attachedFile": {
                    "type": "string"
                },
                "backgroundColor": {
                    "type": "string"
                },
                "backgroundImage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorUserId": {
                    "type": "integer"
                },
                "externalWebLink": {
                    "type": "string"
                },
                "fromTime": {
                    "type": "string"
                },
                "isShareable": {
                    "type": "boolean"
                },
                "relevance": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/DTO.Status"
                },
                "storyID": {
                    "type": "integer"
                },
                "storyName": {
                    "type": "string"
                },
                "toTime": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "DTO.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "DTO.VerifyReq": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
            }
        }
    },
    "securityDefinitions": {
        "GuestAuth": {
            "description": "Just put the guest api key in header.",
            "type": "apiKey",
            "name": "X-Guest-Token",
            "in": "header"
        },
        "JwtAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}